fix(cc-widgets): update-samples
