+++ docs/react-samples/package.json
@@ -13,9 +13,7 @@
     "@babel/preset-env": "^7.25.4",
     "@babel/preset-react": "^7.24.7",
     "@babel/preset-typescript": "^7.25.9",
-    "@r2wc/react-to-web-component": "^2.0.3",
-    "@webex/cc-station-login": "workspace:*",
-    "@webex/cc-user-state": "workspace:*",
+    "@webex/cc-widgets": "workspace:*",
     "babel-loader": "^9.2.1",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.6.3",
