+++ docs/web-component-samples/package.json
@@ -3,8 +3,7 @@
   "version": "1.0.0",
   "main": "dist/index.js",
   "dependencies": {
-    "@webex/cc-station-login": "workspace:*",
-    "@webex/cc-user-state": "workspace:*",
+    "@webex/cc-widgets": "workspace:*",
     "html-webpack-plugin": "^5.6.3",
     "ts-loader": "^9.5.1",
     "typescript": "^5.6.3",
