+++ packages/contact-center/cc-widgets/package.json
@@ -17,6 +17,7 @@
     "test:unit": "jest"
   },
   "dependencies": {
+    "@r2wc/react-to-web-component": "2.0.3",
     "@webex/cc-station-login": "workspace:*",
     "@webex/cc-store": "workspace:*",
     "@webex/cc-user-state": "workspace:*"
