+++ packages/contact-center/station-login/src/index.ts
@@ -1,3 +1,3 @@
-import {StationLogin, WebStationLogin} from './station-login';
+import {StationLogin} from './station-login';
 
-export {StationLogin, WebStationLogin};
+export {StationLogin};
