+++ packages/contact-center/station-login/src/station-login/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import r2wc from '@r2wc/react-to-web-component';
 import store from '@webex/cc-store';
 import {observer} from 'mobx-react';
 
 import StationLoginPresentational from './station-login.presentational';
 import {useStationLogin} from '../helper';
-import { StationLoginProps } from './station-login.types';
+import {StationLoginProps} from './station-login.types';
 
 const StationLogin: React.FunctionComponent<StationLoginProps> = observer(({onLogin, onLogout}) => {
   const {cc, teams, loginOptions} = store;
@@ -14,20 +13,9 @@ const StationLogin: React.FunctionComponent<StationLoginProps> = observer(({onLo
   const props = {
     ...result,
     teams,
-    loginOptions
+    loginOptions,
   };
   return <StationLoginPresentational {...props} />;
 });
 
-const WebStationLogin = r2wc(StationLogin, {
-  props: {
-    onLogin: "function",
-    onLogout: "function"
-  }
-});
-
-if (!customElements.get('widget-cc-station-login')) {
-  customElements.define('widget-cc-station-login', WebStationLogin);
-}
-
-export {StationLogin, WebStationLogin};
+export {StationLogin};
