+++ packages/contact-center/user-state/package.json
@@ -17,7 +17,6 @@
     "test:unit": "jest"
   },
   "dependencies": {
-    "@r2wc/react-to-web-component": "2.0.3",
     "@webex/cc-store": "workspace:*",
     "mobx-react": "9.1.1",
     "react": "18.3.1",
